First Successful Decomposition
First time the test did a successful full disassembling of a mods global.res files including a successful comparison to the ancestor mod (extract only the differences).

The Disassembling process however is not yet finished as the ancestor mod would first need to be reassembled before we can compare it with the child mod. The Reassembling process has not been scripted yet, so in my tests i'm only coparing to __BASE__ which always has a full decoded.txt